refactor: 요구 사항 규칙 준수
- Lotto.js Model 외부 생성
