+++ src/Lotto.js
@@ -1,0 +1,3 @@
+import Lotto from './Model/Lotto.js';
+
+export default Lotto;
