+++ src/docs/README.md
@@ -155,21 +155,29 @@
 
 ### 입력값 처리 관련 테스트
 
-- **입력값 공통 유효성 테스트** `CommonValidatorTest`
+- `CommonValidatorTest`
    - 입력값이 빈 값인 경우
    - 입력값에 공백이 있는 경우
 
 ### 클래스 관련 테스트
 
-- **로또 반환, 결제 금액 유효성 테스트** `LottoMachineTest`
+- `LottoMachineTest`
    - 결제 금액이 1000원 단위로 떨어지지 않는 경우
    - 결제 금액이 숫자가 아닌 다른 입력값이 들어왔을 경우
    - 결제 금액에 맞게 로또가 발행되는지에 대한 경우
-- **로또 발행 테스트** `LottoTest`
+- `LottoTest`
    - 로또 번호가 6개가 아닌 경우
    - 로또 번호에 중복값이 있는 경우
    - 로또 번호의 범위가 1 ~ 45 사이에 있지 않은 경우
    - 로또 번호가 오름차순으로 발행되는지에 대한 경우
+- `WinningLottoTest`
+  - `WinningLotto`의 경우 `Lotto`를 상속 받기 때문에 LottoTest와 동일한 테스트 진행 안함
+  - 보너스 번호가 숫자가 아닌 경우
+  - 보너스 번호가 당첨 번호와 중복값이 경우
+  - 보너스 번호의 범위가 1 ~ 45 사이에 있지 않은 경우
+  - `WinningLotto`가 잘 만들어 지는 경우
+
+
 
 ## 📢 추가된 요구사항
